refactor: BottomNav 아이콘 변경, 경로 수정
